refactor: split large model into smaller files
